Check that YottaDB is installed with UTF-8 support before attempting to install YDBAIM in UTF-8 mode
This commit changes install.sh to make sure that YottaDB was installed with UTF-8 support before attempting to install YDBAIM in UTF-8 mode. Previously, the install script would do this, causing a `ydbinstall --aim` command to attempt to install YDBAIM in UTF-8 mode which would fail and produce an error message.

+++ install.sh
@@ -23,8 +23,10 @@ mkdir build && cd build
 $cmakeCommand -DM_UTF8_MODE=0 ..
 make && make install
 cd ..
-rm -rf build
-mkdir build && cd build
-$cmakeCommand -DM_UTF8_MODE=1 ..
-make && make install
-cd ..
+if [ -d "$ydb_dist/utf8" ] | [ -d "$(pkg-config --variable=prefix yottadb)/utf8" ] ; then
+	rm -rf build
+	mkdir build && cd build
+	$cmakeCommand -DM_UTF8_MODE=1 ..
+	make && make install
+	cd ..
+fi
